re-organize tests
