+++ src/grids.rs
@@ -403,338 +403,3 @@ impl CharGrid for BitGrid {
         return (end_col - col, ch);
     }
 }
-
-
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn height_works() {
-        let grid = BitGrid::new(2, 5);
-        assert_eq!(grid.height(), 5);
-    }
-
-    #[test]
-    fn width_works() {
-        let grid = BitGrid::new(2, 5);
-        assert_eq!(grid.width(), 2*64);
-    }
-
-    #[test]
-    fn create_valid_empty_bitgrid() {
-        let height = 11;
-        let width_in_words = 10;
-        let grid = BitGrid::new(width_in_words, height);
-
-        assert_eq!(grid[0][0], 0);
-        assert_eq!(grid[height-1][width_in_words-1], 0);
-
-        for x in 0..height {
-            for y in 0..width_in_words {
-                assert_eq!(grid[x][y], 0);
-            }
-        }
-    }
-
-    #[test]
-    #[should_panic]
-    fn create_bitgrid_with_invalid_dims() {
-        let height = 0;
-        let width_in_words = 0;
-        let _ = BitGrid::new(width_in_words, height);
-    }
-
-    #[test]
-    fn set_cell_bits_within_a_bitgrid() {
-        let height = 10;
-        let width_in_words = 10;
-        let mut grid = BitGrid::new(width_in_words, height);
-
-        for x in 0..height {
-            for y in 0..width_in_words {
-                assert_eq!(grid[x][y], 0);
-            }
-        }
-
-        grid.modify_bits_in_word(height/2, width_in_words/2, 1<<63, BitOperation::Set);
-        assert_eq!(grid[height/2][width_in_words/2] >> 63, 1);
-        
-        grid.modify_bits_in_word(height-1, width_in_words-1, 1<<63, BitOperation::Set);
-        assert_eq!(grid[height-1][width_in_words-1] >> 63, 1);
-    }
-
-    #[test]
-    fn clear_cell_bits_within_a_bitgrid() {
-        let height = 10;
-        let width_in_words = 10;
-        let mut grid = BitGrid::new(width_in_words, height);
-
-        for x in 0..height {
-            for y in 0..width_in_words {
-                grid[x][y] = u64::max_value();
-            }
-        }
-
-        grid.modify_bits_in_word(height/2, width_in_words/2, 1<<63, BitOperation::Clear);
-        assert_eq!(grid[height/2][width_in_words/2] >> 63, 0);
-        
-        grid.modify_bits_in_word(height-1, width_in_words-1, 1<<63, BitOperation::Clear);
-        assert_eq!(grid[height-1][width_in_words-1] >> 63, 0);
-    }
-
-    #[test]
-    fn toggle_cell_bits_within_a_bitgrid() {
-        let height = 10;
-        let width_in_words = 10;
-        let mut grid = BitGrid::new(width_in_words, height);
-
-        for x in 0..height {
-            for y in 0..width_in_words {
-                grid[x][y] = u64::max_value();
-            }
-        }
-
-        grid.modify_bits_in_word(height/2, width_in_words/2, 1<<63, BitOperation::Toggle);
-        assert_eq!(grid[height/2][width_in_words/2] >> 63, 0);
-        
-        grid.modify_bits_in_word(height/2, width_in_words/2, 1<<63, BitOperation::Toggle);
-        assert_eq!(grid[height/2][width_in_words/2] >> 63, 1);
-    }
-
-    #[test]
-    fn fill_region_within_a_bit_grid() {
-        let height = 10;
-        let width_in_words = 10;
-
-        let region1_w = 7;
-        let region1_h = 7;
-        let region2_w = 3;
-        let region2_h = 3;
-        let region3_h = 4;
-        let region3_w = 4;
-
-        let mut grid = BitGrid::new(width_in_words, height);
-        let region1 = Region::new(0, 0, region1_w, region1_h);
-        let region2 = Region::new(0, 0, region2_w, region2_h);
-        let region3 = Region::new(region2_w as isize, region2_h as isize, region3_w, region3_h);
-
-        grid.modify_region(region1, BitOperation::Set);
-
-        for y in 0..region1_w {
-            assert_eq!(grid[y][0], 0xFE00000000000000);
-        }
-
-        grid.modify_region(region2, BitOperation::Clear);
-        for y in 0..region2_w {
-            assert_eq!(grid[y][0], 0x1E00000000000000);
-        }
-
-        grid.modify_region(region3, BitOperation::Toggle);
-        for x in region2_w..region3_w {
-            for y in region2_h..region3_h {
-                assert_eq!(grid[x][y], 0);
-            }
-        }
-    }
-
-    #[test]
-    fn bounding_box_of_empty_bit_grid_is_none() {
-        let grid = BitGrid::new(1, 1);
-        assert_eq!(grid.bounding_box(), None);
-    }
-
-    #[test]
-    fn bounding_box_of_empty_bit_grid_is_none2() {
-        let grid = BitGrid::new(2, 5);
-        assert_eq!(grid.bounding_box(), None);
-    }
-
-    #[test]
-    fn bounding_box_for_single_off() {
-        let grid = Pattern("b!".to_owned()).to_new_bit_grid(1, 1).unwrap();
-        assert_eq!(grid.bounding_box(), None);
-    }
-
-    #[test]
-    fn bounding_box_for_single_on() {
-        let grid = Pattern("o!".to_owned()).to_new_bit_grid(1, 1).unwrap();
-        assert_eq!(grid.bounding_box(), Some(Region::new(0, 0, 1, 1)));
-    }
-
-    #[test]
-    fn bounding_box_for_single_line_complicated() {
-        let grid = Pattern("15b87o!".to_owned()).to_new_bit_grid(15+87, 1).unwrap();
-        assert_eq!(grid.bounding_box(), Some(Region::new(15, 0, 87, 1)));
-    }
-
-    #[test]
-    fn bounding_box_for_single_line_complicated2() {
-        let grid = Pattern("82b87o!".to_owned()).to_new_bit_grid(82+87, 1).unwrap();
-        assert_eq!(grid.bounding_box(), Some(Region::new(82, 0, 87, 1)));
-    }
-
-    #[test]
-    fn bounding_box_for_multi_line_complicated() {
-        let grid = Pattern("4$15b87o!".to_owned()).to_new_bit_grid(15+87, 5).unwrap();
-        assert_eq!(grid.bounding_box(), Some(Region::new(15, 4, 87, 1)));
-    }
-
-    #[test]
-    fn bounding_box_for_multi_line_complicated2() {
-        let grid = Pattern("4$15b87o$10b100o$25b3o!".to_owned()).to_new_bit_grid(110, 7).unwrap();
-        assert_eq!(grid.bounding_box(), Some(Region::new(10, 4, 100, 3)));
-    }
-
-
-    #[test]
-    fn copy_simple() {
-        let grid = Pattern("64o$64o!".to_owned()).to_new_bit_grid(64, 2).unwrap();
-        let mut grid2 = BitGrid::new(1, 2); // 64x2
-        let dst_region = Region::new(0, 0, 32, 3);
-        BitGrid::copy(&grid, &mut grid2, dst_region);
-        assert_eq!(grid2.to_pattern(None).0, "32o$32o!".to_owned());
-    }
-
-    #[test]
-    fn copy_out_of_bounds() {
-        let grid = Pattern("64o$64o!".to_owned()).to_new_bit_grid(64, 2).unwrap();
-        let mut grid2 = BitGrid::new(1, 2); // 64x2
-        let dst_region = Region::new(17, 3, 32, 5);
-        BitGrid::copy(&grid, &mut grid2, dst_region);
-        assert_eq!(grid2.to_pattern(None).0, "!".to_owned());
-    }
-
-    #[test]
-    fn copy_simple2() {
-        let grid = Pattern("64o$64o!".to_owned()).to_new_bit_grid(64, 2).unwrap();
-        let mut grid2 = BitGrid::new(1, 2); // 64x2
-        let dst_region = Region::new(16, 1, 32, 3);
-        BitGrid::copy(&grid, &mut grid2, dst_region);
-        assert_eq!(grid2.to_pattern(None).0, "$16b32o!".to_owned());
-    }
-
-    #[test]
-    fn copy_for_multi_line_complicated() {
-        let grid = Pattern("4$b87o$3b100o$3o!".to_owned()).to_new_bit_grid(110, 7).unwrap();
-        let mut grid2 = BitGrid::new(2, 10); // 128x10
-        let dst_region = Region::new(2, 3, 64, 5);
-        BitGrid::copy(&grid, &mut grid2, dst_region);
-        assert_eq!(grid2.to_pattern(None).0, "7$3b63o!".to_owned());
-    }
-
-    #[test]
-    fn copy_for_multi_line_complicated2() {
-        let grid = Pattern("4$b87o$3b100o$3o!".to_owned()).to_new_bit_grid(110, 7).unwrap();
-        let mut grid2 = BitGrid::new(2, 10); // 128x10
-        let dst_region = Region::new(2, 3, 65, 5);
-        BitGrid::copy(&grid, &mut grid2, dst_region);
-        assert_eq!(grid2.to_pattern(None).0, "7$3b64o!".to_owned());
-    }
-
-    #[test]
-    fn copy_for_multi_line_complicated3() {
-        let grid = Pattern("4$b87o$3b100o$3o!".to_owned()).to_new_bit_grid(110, 7).unwrap();
-        let mut grid2 = BitGrid::new(1, 10); // 64x10
-        let dst_region = Region::new(2, 3, 5, 5);
-        BitGrid::copy(&grid, &mut grid2, dst_region);
-        assert_eq!(grid2.to_pattern(None).0, "7$3b4o!".to_owned());
-    }
-
-
-    #[test]
-    #[should_panic]
-    fn modify_region_with_a_negative_region_panics() {
-        let height = 10;
-        let width_in_words = 10;
-
-        let mut grid = BitGrid::new(width_in_words, height);
-        let region_neg = Region::new(-1, -1, 1, 1);
-        grid.modify_region(region_neg, BitOperation::Set);
-    }
-
-    #[test]
-    fn get_run_for_single_on() {
-        let grid = Pattern("o!".to_owned()).to_new_bit_grid(1, 1).unwrap();
-        assert_eq!(grid.get_run(0, 0, None), (1, 'o'));
-    }
-
-    #[test]
-    fn get_run_for_single_off_then_on() {
-        let grid = Pattern("bo!".to_owned()).to_new_bit_grid(2, 1).unwrap();
-        assert_eq!(grid.get_run(0, 0, None), (1, 'b'));
-    }
-
-    #[test]
-    fn get_run_for_single_off() {
-        let grid = Pattern("b!".to_owned()).to_new_bit_grid(1, 1).unwrap();
-        assert_eq!(grid.get_run(0, 0, None), (64, 'b'));
-    }
-
-    #[test]
-    fn get_run_nonzero_col_for_single_off() {
-        let grid = Pattern("b!".to_owned()).to_new_bit_grid(1, 1).unwrap();
-        assert_eq!(grid.get_run(23, 0, None), (64 - 23, 'b'));
-    }
-
-    #[test]
-    fn get_run_nonzero_col_complicated() {
-        let grid = Pattern("15b87o!".to_owned()).to_new_bit_grid(15+87, 1).unwrap();
-        assert_eq!(grid.get_run(15, 0, None), (87, 'o'));
-    }
-
-    #[test]
-    fn get_run_nonzero_col_multiline_complicated() {
-        let grid = Pattern("3$15b87o!".to_owned()).to_new_bit_grid(15+87, 4).unwrap();
-        assert_eq!(grid.get_run(15, 3, None), (87, 'o'));
-    }
-
-
-
-    #[test]
-    fn to_pattern_simple() {
-        let original_pattern = Pattern("bo!".to_owned());
-        let grid = original_pattern.to_new_bit_grid(2, 1).unwrap();
-        let pattern = grid.to_pattern(None);
-        assert_eq!(pattern, original_pattern);
-    }
-
-    #[test]
-    fn to_pattern_nonzero_col_3empty_then_complicated() {
-        let original_pattern = Pattern("3$15b87o!".to_owned());
-        let grid = original_pattern.to_new_bit_grid(15+87, 4).unwrap();
-        let pattern = grid.to_pattern(None);
-        assert_eq!(pattern, original_pattern);
-    }
-
-    #[test]
-    fn to_pattern_nonzero_col_veryverycomplicated() {
-        let original_pattern = Pattern("b2o23b2o21b$b2o23bo22b$24bobo22b$15b2o7b2o23b$2o13bobo31b$2o13bob2o30b$16b2o31b$16bo32b$44b2o3b$16bo27b2o3b$16b2o31b$2o13bob2o13bo3bo12b$2o13bobo13bo5bo7b2o2b$15b2o14bo13b2o2b$31b2o3bo12b$b2o30b3o13b$b2o46b$33b3o13b$31b2o3bo12b$31bo13b2o2b$31bo5bo7b2o2b$32bo3bo12b2$44b2o3b$44b2o3b5$37b2o10b$37bobo7b2o$39bo7b2o$37b3o9b$22bobo24b$21b3o25b$21b3o25b$21bo15b3o9b$25bobo11bo9b$21b2o4bo9bobo9b$16b2o4bo3b2o9b2o10b$15bobo6bo24b$15bo33b$14b2o!".to_owned());
-        let grid = original_pattern.to_new_bit_grid(49, 43).unwrap();
-        let pattern = grid.to_pattern(None);
-        assert_eq!(pattern.0.as_str(), "b2o23b2o$b2o23bo$24bobo$15b2o7b2o$2o13bobo$2o13bob2o$16b2o$16bo$44b2o$\r\n16bo27b2o$16b2o$2o13bob2o13bo3bo$2o13bobo13bo5bo7b2o$15b2o14bo13b2o$\r\n31b2o3bo$b2o30b3o$b2o$33b3o$31b2o3bo$31bo13b2o$31bo5bo7b2o$32bo3bo2$\r\n44b2o$44b2o5$37b2o$37bobo7b2o$39bo7b2o$37b3o$22bobo$21b3o$21b3o$21bo\r\n15b3o$25bobo11bo$21b2o4bo9bobo$16b2o4bo3b2o9b2o$15bobo6bo$15bo$14b2o!");
-        for line in pattern.0.as_str().lines() {
-            assert!(line.len() <= 70);
-        }
-    }
-
-    #[test]
-    fn get_run_empty() {
-        let grid = BitGrid::new(1,1);
-        let pattern = grid.to_pattern(None);
-        assert_eq!(pattern.0.as_str(), "!");
-    }
-
-    #[test]
-    fn clear_all_ones_really_is_clear() {
-        let pat = Pattern("64o$64o$64o$64o$64o$64o$64o$64o$64o$64o$64o$64o!".to_owned());
-        let mut grid = pat.to_new_bit_grid(64, 12).unwrap();
-        grid.clear();
-        for row in &grid.0 {
-            for col_idx in 0..row.len() {
-                assert_eq!(row[col_idx], 0);
-            }
-        }
-    }
-}
