+++ src/rle.rs
@@ -31,12 +31,12 @@ pub struct Pattern(pub String);
 #[derive(Debug, PartialEq, Clone)]
 pub struct PatternFile {
     pub comment_lines: Vec<String>,
-    header_line: HeaderLine,
+    pub header_line: HeaderLine,
     pub pattern: Pattern,
 }
 
 #[derive(Debug, PartialEq, Clone)]
-struct HeaderLine {
+pub struct HeaderLine {
     pub x: usize, // width (cols)
     pub y: usize, // height (rows)
     pub rule: Option<String>,
@@ -241,131 +241,5 @@ impl Pattern {
             return Err("Premature termination".to_owned());
         }
         Ok(())
-    }
-}
-
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    // Glider gun
-    #[test]
-    fn one_line_parsing_works1() {
-        let gun = Pattern("24bo$22bobo$12b2o6b2o12b2o$11bo3bo4b2o12b2o$2o8bo5bo3b2o$2o8bo3bob2o4bobo$10bo5bo7bo$11bo3bo$12b2o!".to_owned()).to_new_bit_grid(36, 9).unwrap();
-        assert_eq!(gun[0][0], 0b0000000000000000000000001000000000000000000000000000000000000000);
-        assert_eq!(gun[1][0], 0b0000000000000000000000101000000000000000000000000000000000000000);
-        assert_eq!(gun[2][0], 0b0000000000001100000011000000000000110000000000000000000000000000);
-        assert_eq!(gun[3][0], 0b0000000000010001000011000000000000110000000000000000000000000000);
-        assert_eq!(gun[4][0], 0b1100000000100000100011000000000000000000000000000000000000000000);
-        assert_eq!(gun[5][0], 0b1100000000100010110000101000000000000000000000000000000000000000);
-        assert_eq!(gun[6][0], 0b0000000000100000100000001000000000000000000000000000000000000000);
-        assert_eq!(gun[7][0], 0b0000000000010001000000000000000000000000000000000000000000000000);
-        assert_eq!(gun[8][0], 0b0000000000001100000000000000000000000000000000000000000000000000);
-    }
-
-
-    // Glider gun with line break
-    #[test]
-    fn multi_line_parsing_works1() {
-        let gun = Pattern("24bo$22bobo$12b2o6b2o\r\n12b2o$\r\n11bo3bo4b2o12b2o$2o8b\ro5bo3b2o$2o8bo3bob2o4b\nobo$10bo5bo7bo$11bo3bo$12b2o!".to_owned()).to_new_bit_grid(36, 9).unwrap();
-        assert_eq!(gun[0][0], 0b0000000000000000000000001000000000000000000000000000000000000000);
-        assert_eq!(gun[1][0], 0b0000000000000000000000101000000000000000000000000000000000000000);
-        assert_eq!(gun[2][0], 0b0000000000001100000011000000000000110000000000000000000000000000);
-        assert_eq!(gun[3][0], 0b0000000000010001000011000000000000110000000000000000000000000000);
-        assert_eq!(gun[4][0], 0b1100000000100000100011000000000000000000000000000000000000000000);
-        assert_eq!(gun[5][0], 0b1100000000100010110000101000000000000000000000000000000000000000);
-        assert_eq!(gun[6][0], 0b0000000000100000100000001000000000000000000000000000000000000000);
-        assert_eq!(gun[7][0], 0b0000000000010001000000000000000000000000000000000000000000000000);
-        assert_eq!(gun[8][0], 0b0000000000001100000000000000000000000000000000000000000000000000);
-    }
-
-    #[test]
-    fn parsing_with_new_row_rle_works() {
-        let pattern = Pattern("27bo$28bo$29bo$28bo$27bo$29b3o20$oo$bbo$bbo$3b4o!".to_owned()).to_new_bit_grid(32, 29).unwrap();
-        assert_eq!(pattern.0[0..=5],
-           [[0b0000000000000000000000000001000000000000000000000000000000000000],
-            [0b0000000000000000000000000000100000000000000000000000000000000000],
-            [0b0000000000000000000000000000010000000000000000000000000000000000],
-            [0b0000000000000000000000000000100000000000000000000000000000000000],
-            [0b0000000000000000000000000001000000000000000000000000000000000000],
-            [0b0000000000000000000000000000011100000000000000000000000000000000]]);
-        for row in 6..=24 {
-            assert_eq!(pattern.0[row][0], 0);
-        }
-        assert_eq!(pattern.0[25..=28],
-            [[0b1100000000000000000000000000000000000000000000000000000000000000],
-             [0b0010000000000000000000000000000000000000000000000000000000000000],
-             [0b0010000000000000000000000000000000000000000000000000000000000000],
-             [0b0001111000000000000000000000000000000000000000000000000000000000]]);
-    }
-
-
-    #[test]
-    fn parse_whole_file_works() {
-        let pat: PatternFile = PatternFile::from_str("#N Gosper glider gun\n#C This was the first gun discovered.\n#C As its name suggests, it was discovered by Bill Gosper.\nx = 36, y = 9, rule = B3/S23\n24bo$22bobo$12b2o6b2o12b2o$11bo3bo4b2o12b2o$2o8bo5bo3b2o$2o8bo3bob2o4b\nobo$10bo5bo7bo$11bo3bo$12b2o!\n").unwrap();
-        assert_eq!(pat.comment_lines.len(), 3);
-        assert_eq!(pat.header_line, HeaderLine{x: 36, y: 9, rule: Some("B3/S23".to_owned())});
-        assert_eq!(pat.pattern.0, "24bo$22bobo$12b2o6b2o12b2o$11bo3bo4b2o12b2o$2o8bo5bo3b2o$2o8bo3bob2o4bobo$10bo5bo7bo$11bo3bo$12b2o!");
-    }
-
-    #[test]
-    fn parse_whole_file_works_with_crap_at_the_end() {
-        let pat: PatternFile = PatternFile::from_str("#N Gosper glider gun\n#C This was the first gun discovered.\n#C As its name suggests, it was discovered by Bill Gosper.\nx = 36, y = 9, rule = B3/S23\n24bo$22bobo$12b2o6b2o12b2o$11bo3bo4b2o12b2o$2o8bo5bo3b2o$2o8bo3bob2o4b\nobo$10bo5bo7bo$11bo3bo$12b2o!blah\n\nyaddayadda\n").unwrap();
-        assert_eq!(pat.comment_lines.len(), 3);
-        assert_eq!(pat.header_line, HeaderLine{x: 36, y: 9, rule: Some("B3/S23".to_owned())});
-        assert_eq!(pat.pattern.0, "24bo$22bobo$12b2o6b2o12b2o$11bo3bo4b2o12b2o$2o8bo5bo3b2o$2o8bo3bob2o4bobo$10bo5bo7bo$11bo3bo$12b2o!");
-    }
-
-    #[test]
-    fn parse_whole_file_works_vacuum_cleaner() {
-        let pat: PatternFile = PatternFile::from_str("#N Vacuum (gun)\r\n#O Dieter Leithner\r\n#C A true period 46 double-barreled gun found on February 21, 1997.\r\n#C www.conwaylife.com/wiki/index.php?title=Vacuum_(gun)\r\nx = 49, y = 43, rule = b3/s23\r\nb2o23b2o21b$b2o23bo22b$24bobo22b$15b2o7b2o23b$2o13bobo31b$2o13bob2o30b\r\n$16b2o31b$16bo32b$44b2o3b$16bo27b2o3b$16b2o31b$2o13bob2o13bo3bo12b$2o\r\n13bobo13bo5bo7b2o2b$15b2o14bo13b2o2b$31b2o3bo12b$b2o30b3o13b$b2o46b$\r\n33b3o13b$31b2o3bo12b$31bo13b2o2b$31bo5bo7b2o2b$32bo3bo12b2$44b2o3b$44b\r\n2o3b5$37b2o10b$37bobo7b2o$39bo7b2o$37b3o9b$22bobo24b$21b3o25b$21b3o25b\r\n$21bo15b3o9b$25bobo11bo9b$21b2o4bo9bobo9b$16b2o4bo3b2o9b2o10b$15bobo6b\r\no24b$15bo33b$14b2o!").unwrap();
-        assert_eq!(pat.comment_lines.len(), 4);
-        assert_eq!(pat.header_line, HeaderLine{x: 49, y: 43, rule: Some("B3/S23".to_owned().to_lowercase())});
-        assert_eq!(pat.pattern.0, "b2o23b2o21b$b2o23bo22b$24bobo22b$15b2o7b2o23b$2o13bobo31b$2o13bob2o30b$16b2o31b$16bo32b$44b2o3b$16bo27b2o3b$16b2o31b$2o13bob2o13bo3bo12b$2o13bobo13bo5bo7b2o2b$15b2o14bo13b2o2b$31b2o3bo12b$b2o30b3o13b$b2o46b$33b3o13b$31b2o3bo12b$31bo13b2o2b$31bo5bo7b2o2b$32bo3bo12b2$44b2o3b$44b2o3b5$37b2o10b$37bobo7b2o$39bo7b2o$37b3o9b$22bobo24b$21b3o25b$21b3o25b$21bo15b3o9b$25bobo11bo9b$21b2o4bo9bobo9b$16b2o4bo3b2o9b2o10b$15bobo6bo24b$15bo33b$14b2o!");
-        assert_eq!(pat.to_new_bit_grid().unwrap(), BitGrid(vec![
-            vec![0b0110000000000000000000000011000000000000000000000000000000000000],
-            vec![0b0110000000000000000000000010000000000000000000000000000000000000],
-            vec![0b0000000000000000000000001010000000000000000000000000000000000000],
-            vec![0b0000000000000001100000001100000000000000000000000000000000000000],
-            vec![0b1100000000000001010000000000000000000000000000000000000000000000],
-            vec![0b1100000000000001011000000000000000000000000000000000000000000000],
-            vec![0b0000000000000000110000000000000000000000000000000000000000000000],
-            vec![0b0000000000000000100000000000000000000000000000000000000000000000],
-            vec![0b0000000000000000000000000000000000000000000011000000000000000000],
-            vec![0b0000000000000000100000000000000000000000000011000000000000000000],
-            vec![0b0000000000000000110000000000000000000000000000000000000000000000],
-            vec![0b1100000000000001011000000000000010001000000000000000000000000000],
-            vec![0b1100000000000001010000000000000100000100000001100000000000000000],
-            vec![0b0000000000000001100000000000000100000000000001100000000000000000],
-            vec![0b0000000000000000000000000000000110001000000000000000000000000000],
-            vec![0b0110000000000000000000000000000001110000000000000000000000000000],
-            vec![0b0110000000000000000000000000000000000000000000000000000000000000],
-            vec![0b0000000000000000000000000000000001110000000000000000000000000000],
-            vec![0b0000000000000000000000000000000110001000000000000000000000000000],
-            vec![0b0000000000000000000000000000000100000000000001100000000000000000],
-            vec![0b0000000000000000000000000000000100000100000001100000000000000000],
-            vec![0b0000000000000000000000000000000010001000000000000000000000000000],
-            vec![0b0000000000000000000000000000000000000000000000000000000000000000],
-            vec![0b0000000000000000000000000000000000000000000011000000000000000000],
-            vec![0b0000000000000000000000000000000000000000000011000000000000000000],
-            vec![0b0000000000000000000000000000000000000000000000000000000000000000],
-            vec![0b0000000000000000000000000000000000000000000000000000000000000000],
-            vec![0b0000000000000000000000000000000000000000000000000000000000000000],
-            vec![0b0000000000000000000000000000000000000000000000000000000000000000],
-            vec![0b0000000000000000000000000000000000000110000000000000000000000000],
-            vec![0b0000000000000000000000000000000000000101000000011000000000000000],
-            vec![0b0000000000000000000000000000000000000001000000011000000000000000],
-            vec![0b0000000000000000000000000000000000000111000000000000000000000000],
-            vec![0b0000000000000000000000101000000000000000000000000000000000000000],
-            vec![0b0000000000000000000001110000000000000000000000000000000000000000],
-            vec![0b0000000000000000000001110000000000000000000000000000000000000000],
-            vec![0b0000000000000000000001000000000000000111000000000000000000000000],
-            vec![0b0000000000000000000000000101000000000001000000000000000000000000],
-            vec![0b0000000000000000000001100001000000000101000000000000000000000000],
-            vec![0b0000000000000000110000100011000000000110000000000000000000000000],
-            vec![0b0000000000000001010000001000000000000000000000000000000000000000],
-            vec![0b0000000000000001000000000000000000000000000000000000000000000000],
-            vec![0b0000000000000011000000000000000000000000000000000000000000000000]]));
     }
 }
